sandbox: Move `InMemoryPackageStore` into its own `packages` package.
`packages` package. What a tongue twister.
