+++ ledger/sandbox/src/main/scala/com/digitalasset/platform/packages/InMemoryPackageStore.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2020 The DAML Authors. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.digitalasset.platform.apiserver
+package com.digitalasset.platform.packages
 
 import java.io.File
 import java.time.Instant
